+++ NEWS.md
@@ -1,6 +1,6 @@
 # version 0.8.0 
 
-* get_all --- iterate through the results and get all supported for various functions.
+* get_all --- iterate through the results and get all supported for various functions. supported for yt_search(). prints removed from yt_search()
 
 # version 0.7.0